chore: fixes

+++ src/commands/helpers/createConfig.ts
@@ -1,5 +1,6 @@
 import fs from 'node:fs'
 import path from 'node:path'
+import { findConfig } from './loadConfig'
 
 const CONFIG_FILE_NAME = 'next-serverless.config.js'
 const CONFIG_TEMPLATE = `/**
@@ -17,8 +18,9 @@ module.exports = config
 
 export const createConfigFile = () => {
   const configFilePath = path.resolve(process.cwd(), CONFIG_FILE_NAME)
+  const serverConfig = findConfig(process.cwd())
 
-  if (!fs.existsSync(configFilePath)) {
+  if (!serverConfig && !fs.existsSync(configFilePath)) {
     fs.writeFileSync(configFilePath, CONFIG_TEMPLATE, 'utf-8')
     console.log(`Created sample configuration file at ${configFilePath}`)
   } else {
